responsive layout/sizes for ui redesign
