--- src/components/ship/ship_mobile.module.css
@@ -1,155 +1,0 @@
-.container {
-    display: flex;
-    height: auto;
-}
-
-.shipContainer {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.topDiv {
-    width: 100%;
-}
-
-.shipUI {
-    text-align: center;
-    font-weight: 600;
-    display: flex;
-    justify-content: space-evenly;
-}
-
-.playerName {
-    border-bottom: 1px solid;
-    width: 100%;
-    font-size: 1.25rem;
-    font-weight: 600;
-}
-
-.player1Name {
-    text-align: right;
-}
-
-.player2Name {
-    text-align: left;
-}
-
-.statusEffectsContainer {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    min-width: 24px;
-}
-
-.statusIcon {
-    padding: 2px;
-    margin: 4px 0;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: rgba(51, 51, 51, 0.8);
-}
-
-.effectsContainer {
-    position: absolute;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.effect {
-    opacity: 0;
-    position: fixed;
-}
-
-.resources {
-    display: flex;
-    /* background-color: rgba(51, 51, 51, 0.6); */
-}
-
-.resource {
-    display: flex;
-    /* width: 100%; */
-    background-color: rgba(51, 51, 51, 0.6);
-    margin: 5px;
-}
-
-.prodContainer {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-.prodDiv {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 1.5rem;
-    font-weight: 600;
-}
-
-.imgDiv {
-    width: 30px;
-    height: 30px;
-}
-
-.resourceImgContainer {
-    font-size: 1.25rem;
-    font-weight: 600;
-    position: relative;
-    border-radius: 50%;
-    text-align: center;
-    display: flex;
-    justify-content: center;
-    align-items: flex-start;
-    padding-top: 10px;
-    padding-right: 6px;
-}
-
-.resourceCounter {
-    position: absolute;
-    border-radius: 50%;
-    width: 22px;
-    height: 22px;
-    text-align: center;
-    transform: translate(8px, 18px);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.resourceCount {
-    font-size: 1rem;
-}
-
-.defenseImgDiv {
-    background-image: url('../../../public/images/defense.png');
-}
-.powerImgDiv {
-    background-image: url('../../../public/images/power.png');
-}
-.offenseImgDiv {
-    background-image: url('../../../public/images/offense.png');
-}
-
-.shipImage {
-    height: 150px;
-    width: 200px;
-}
-
-.shipImageDiv {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-}
-
-@media only screen and (max-height: 550) {
-    .shipImage {
-        height: auto;
-        width: auto;
-        object-fit: contain;
-    }
-}
